refactor: merge default partial encoders/decoders
